+++ src/main/java/net/geforcemods/securitycraft/blocks/UsernameLoggerBlock.java
@@ -6,10 +6,12 @@ import net.geforcemods.securitycraft.blockentities.UsernameLoggerBlockEntity;
 import net.geforcemods.securitycraft.util.LevelUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.core.Direction.Axis;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
 import net.minecraft.world.level.block.Block;
@@ -38,6 +40,11 @@ public class UsernameLoggerBlock extends DisguisableBlock {
 			ClientHandler.displayUsernameLoggerScreen(level, pos);
 
 		return InteractionResult.SUCCESS;
+	}
+
+	@Override
+	public boolean canConnectRedstone(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {
+		return direction != null && direction.getAxis() != Axis.Y;
 	}
 
 	@Override