+++ Changelog.md
@@ -3,6 +3,7 @@
 - New: Support for FTB Teams when having team ownership enabled in the config
 - Fix: Turning off camera rotation leads to extreme stuttering of the camera
 - Fix: Shift-clicking a potion in the brewing stand screen puts the potion into the wrong slot
+- Fix: Redstone does not automatically connect to the Alarm, Projector, and Username Logger
 
 --------------------------Changelog for v1.9.5 of SecurityCraft--------------------------